WIP removing stuff that we shouldn't keeep around
