+++ workspaces/cli-server/package.json
@@ -35,7 +35,6 @@
     "find-process": "^1.4.3",
     "fs-extra": "^9.0.0",
     "get-port": "^5.1.0",
-    "http-proxy-middleware": "^1.0.4",
     "is-port-reachable": "^3.0.0",
     "jwt-decode": "2.2.0",
     "lodash.sortby": "^4.7.0",
@@ -55,7 +54,6 @@
     "@types/express": "^4.17.2",
     "@types/find-process": "^1.2.0",
     "@types/fs-extra": "^8.0.1",
-    "@types/http-proxy-middleware": "^0.19.3",
     "@types/lodash.sortby": "^4.7.6",
     "@types/node": "^13.1.6",
     "@types/node-fetch": "^2.5.5",
