+++ workspaces/cli-server/src/server.ts
@@ -16,7 +16,6 @@ import {
   makeRouter,
 } from './routers/spec-router';
 import { basePath } from '@useoptic/ui-v2';
-import { createProxyMiddleware } from 'http-proxy-middleware';
 import { TrackingEventBase } from '@useoptic/analytics/lib/interfaces/TrackingEventBase';
 import { analyticsEvents, trackWithApiName } from './analytics';
 import cors from 'cors';
