+++ workspaces/cli-server/src/sessions.ts
@@ -1,17 +1,3 @@
-import {
-  getPathsRelativeToCwd,
-  readApiConfig,
-  IApiCliConfig,
-} from '@useoptic/cli-config';
-import {
-  createDiff,
-  Diff,
-  DiffQueries,
-  ProgressStream as DiffProgressStream,
-} from './diffs';
-import { OnDemandDiff } from './diffs/on-demand';
-import * as Uuid from 'uuid';
-
 export class SessionsManager {
   private sessions: Session[] = [];
 
@@ -45,109 +31,8 @@ export class SessionsManager {
 
 export class Session {
   constructor(readonly id: string, readonly path: string) {}
-  diffs: SessionDiffs | null = null;
-  async start() {
-    const resolvedPaths = await getPathsRelativeToCwd(this.path);
-    this.diffs = new SessionDiffs(
-      resolvedPaths.configPath,
-      resolvedPaths.opticIgnorePath,
-      resolvedPaths.capturesPath,
-      resolvedPaths.specStorePath,
-      resolvedPaths.specDirPath
-    );
-  }
 
-  async diffCapture(
-    captureId: string,
-    events?: Array<{ [key: string]: any }>,
-    endpoints?: Array<{ pathId: string; method: string }>
-  ) {
-    if (!this.diffs)
-      throw new Error(
-        'Session must have been started before it can diff a capture'
-      );
-
-    return this.diffs.startDiff(captureId, events, endpoints);
-  }
-
-  diffProgress(diffId: string) {
-    if (!this.diffs)
-      throw new Error(
-        'Session must have been started before diffs can be accessed'
-      );
-
-    return this.diffs.progress(diffId);
-  }
-
-  diffQueries(diffId: string) {
-    if (!this.diffs)
-      throw new Error(
-        'Session must have been started before diffs can be accessed'
-      );
-
-    return this.diffs.queries(diffId);
-  }
+  async start() {}
 
   async stop() {}
-}
-
-class SessionDiffs {
-  private diffsById: Map<string, Diff> = new Map();
-  private activeDiffsByCaptureId: Map<string, Diff> = new Map();
-
-  constructor(
-    readonly configPath: string,
-    readonly opticIgnorePath: string,
-    readonly capturesPath: string,
-    readonly specPath: string,
-    readonly specDirPath: string
-  ) {}
-
-  async startDiff(
-    captureId: string,
-    events?: Array<{ [key: string]: any }>,
-    endpoints?: Array<{ pathId: string; method: string }>
-  ): Promise<string> {
-    const diffId = Uuid.v4();
-    const newDiff = createDiff(OnDemandDiff, {
-      captureId,
-      opticIgnorePath: this.opticIgnorePath,
-      configPath: this.configPath,
-      captureBaseDirectory: this.capturesPath,
-      diffId,
-      events,
-      endpoints,
-      specPath: this.specPath,
-      specDirPath: this.specDirPath,
-    });
-    this.activeDiffsByCaptureId.set(captureId, newDiff);
-    this.diffsById.set(diffId, newDiff);
-
-    newDiff.events.once('finish', () => {
-      this.activeDiffsByCaptureId.delete(captureId);
-    });
-    newDiff.events.once('error', (err) => {
-      debugger;
-      console.error(err);
-      throw err;
-    });
-
-    await newDiff.start();
-
-    return diffId;
-  }
-
-  progress(diffId: string): DiffProgressStream | undefined {
-    const diff = this.diffsById.get(diffId);
-    if (!diff) return;
-
-    return diff.progress();
-  }
-
-  queries(diffId: string): DiffQueries | undefined {
-    const diff = this.diffsById.get(diffId);
-    if (!diff) return;
-
-    return diff.queries();
-  }
 }
