+++ workspaces/cli-server/src/tasks/on-demand-initial-body-rust.ts
@@ -5,7 +5,6 @@ import {
 } from '@useoptic/cli-shared/build/diffs/initial-bodies-worker-rust';
 
 export { LearnedBodies };
-//@dev: follow this pattern for diff worker
 export class OnDemandInitialBodyRust {
   private worker: InitialBodiesWorkerRust;
