--- workspaces/cli-server/src/tasks/on-demand-trail-values.ts
@@ -1,155 +1,0 @@
-import { EventEmitter } from 'events';
-import { runManagedScriptByName } from '@useoptic/cli-scripts';
-import { ChildProcess } from 'child_process';
-import fs from 'fs-extra';
-import {
-  IValueAffordanceSerializationWithCounter,
-  IValueAffordanceSerializationWithCounterGroupedByDiffHash,
-} from '@useoptic/cli-shared/build/diffs/initial-types';
-import {
-  getTrailWorkerOutputPaths,
-  ITrailValuesEmitterWorker,
-} from '@useoptic/cli-shared/build/diffs/trail-values-worker';
-
-export interface ITrailValuesConfig {
-  pathId: string;
-  method: string;
-  captureId: string;
-  events: any;
-  serializedDiffs: string;
-  captureBaseDirectory: string;
-}
-
-export class OnDemandTrailValues {
-  public readonly events: EventEmitter = new EventEmitter();
-  private child!: ChildProcess;
-  public readonly id: string;
-  private readonly config: ITrailValuesConfig;
-  private finished: boolean = false;
-
-  constructor(config: ITrailValuesConfig) {
-    this.id = config.method + config.method + config.captureId;
-    this.config = config;
-  }
-
-  private lastProgress: {
-    hasMoreInteractions: boolean;
-    results: IValueAffordanceSerializationWithCounterGroupedByDiffHash;
-  } | null = null;
-
-  async run(): Promise<
-    IValueAffordanceSerializationWithCounterGroupedByDiffHash
-  > {
-    const { config } = this;
-
-    const outputPaths = this.paths();
-
-    const scriptConfig: ITrailValuesEmitterWorker = {
-      captureId: config.captureId,
-      pathId: config.pathId,
-      method: config.method,
-      specFilePath: outputPaths.events,
-      serializedDiffs: config.serializedDiffs,
-      captureBaseDirectory: config.captureBaseDirectory,
-    };
-
-    await fs.ensureDir(outputPaths.base);
-
-    await Promise.all([fs.writeJson(outputPaths.events, config.events)]);
-
-    console.log(JSON.stringify(scriptConfig));
-
-    const child: ChildProcess = runManagedScriptByName(
-      'emit-learn-trail-value',
-      JSON.stringify(scriptConfig)
-    );
-
-    const onMessage = (x: any) => {
-      if (x.type && x.type === 'progress') {
-        if (!x.data.hasMoreInteractions) {
-          this.finished = true;
-          this.events.emit('completed', x.data);
-        }
-      }
-      this.events.emit(x.type, x.data);
-    };
-    const onError = (err: Error) => {
-      cleanup();
-      this.events.emit('error', err);
-    };
-    const onExit = (code: number, signal: string | null) => {
-      cleanup();
-      if (code !== 0) {
-        // @TODO: wonder how we'll ever find out about this happening.
-        console.error(
-          `On Demand Trail Value Worker exited with non-zero exit code ${code}`
-        );
-      } else {
-        this.finished = true;
-        this.events.emit('finish');
-      }
-    };
-
-    function cleanup() {
-      child.removeListener('message', onMessage);
-      child.removeListener('error', onError);
-      child.removeListener('exit', onExit);
-    }
-
-    child.on('message', onMessage);
-    child.once('error', onError);
-    child.once('exit', onExit);
-
-    this.child = child;
-
-    const completedPromise: Promise<IValueAffordanceSerializationWithCounterGroupedByDiffHash> = new Promise(
-      (resolve) => {
-        this.events.once('completed', onCompleted);
-        function onCompleted(data: any) {
-          cleanup();
-          child.kill();
-          resolve(data.results);
-        }
-        const cleanup = () => {
-          this.events.removeListener('completed', onCompleted);
-        };
-      }
-    );
-
-    const startedPromise = await new Promise(async (resolve, reject) => {
-      function onErr(err: Error) {
-        cleanup();
-        reject(err);
-      }
-      function onProgress(data: any) {
-        cleanup();
-        resolve();
-      }
-      function onFinish() {
-        cleanup();
-        resolve();
-      }
-      const cleanup = () => {
-        this.events.removeListener('progress', onProgress);
-        this.events.removeListener('error', onErr);
-        this.events.removeListener('finish', onFinish);
-      };
-
-      this.events.once('progress', onProgress);
-      this.events.once('error', onErr);
-    });
-
-    await startedPromise;
-    return await completedPromise;
-  }
-
-  private paths(): { base: string; events: string } {
-    return getTrailWorkerOutputPaths(this.config);
-  }
-
-  async stop() {
-    if (this.child) {
-      this.child.kill('SIGTERM');
-    }
-  }
-}
