+++ workspaces/cli-shared/src/diffs/trail-values-worker-rust.ts
@@ -3,7 +3,7 @@ import fs from 'fs-extra';
 import { learnShapeDiffAffordances } from '@useoptic/diff-engine';
 import { Streams } from '@useoptic/diff-engine-wasm';
 import { ShapeDiffAffordances } from '@useoptic/diff-engine-wasm/lib/streams/learning-results/shape-diff-affordances';
-import { getDiffOutputPaths } from './diff-worker-rust';
+import { getDiffOutputPaths } from './interaction-diff-worker-rust';
 import path from 'path';
 
 export interface ShapeDiffAffordancesConfig {
